Renaming the admin middleware
